requirements.txt
gradio_imageslider
gradio==4.29.0
matplotlib
opencv-python
torch
torchvision
pillow-heif
pillow